requirements.txt
boxsdk[jwt]>=3.9.1
google-api-python-client>=2.136.0
google-auth>=2.35.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.1
google-cloud-vision>=3.7.4
google-cloud-videointelligence>=2.11.0
google-resumable-media>=2.7.2
httplib2>=0.22.0
fastapi>=0.115.2
uvicorn[standard]>=0.30.6
pydantic>=2.8.2
python-dotenv>=1.0.1
PyYAML>=6.0.2
tenacity>=9.0.0
Jinja2>=3.1.4
requests>=2.32.3
rich>=13.8.1

